chore(ai-assistant): implement premium status checks and Suspense for loading state
- Introduced a new `fetchPremiumStatus` function to determine user premium status using localStorage and Firebase custom claims.
- Refactored `AIAssistantPage` to utilize Suspense for loading premium status, enhancing user experience with a loading skeleton.
- Created `AIAssistantContent` component to separate logic and improve maintainability.
- Updated mobile dashboard to reflect premium features and ensure consistent layout.
- Enhanced account deletion process in settings with improved error handling and user feedback.
