+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cannafriend",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "private": true,
   "scripts": {
     "build": "next build",
